+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+# 2022-04-05 - v.5.8.3
+- Apptentive Android SDK: 5.8.3
+- Apptentive iOS SDK: 5.3.4
+
+# 2022-01-31 - v5.8.0-rc.7
+- Apptentive Android SDK: 5.8.0
+- Apptentive iOS SDK: 5.3.4
+- Remove deprecated login feature
+- Update jcenter to mavenCentral
+
+# 2021-12-10 - v5.7.1-rc.6+hotfix.1
+- Fix merge issue with duplicated methods
+- Fix FlutterFire warning in iOS example app
+
+# 2021-12-09 - v5.7.1-rc.6
+- Implement setPushNotificationIntegration()
+- Ask for push permissions when Message Center is used on iOS
+
+# 2021-11-10 - v5.7.1-rc.5
+- Added and fixed callback listeners on Android and iOS
+- Implemented getUnreadMessageCount()
+
+# 2021-10-14 - v5.7.1-rc.4
+- Added distribution information
+- Fixed Android log level bug
+
+# 2021-10-13 - v5.7.1-rc.3
+- Fixed iOS registration checks and logs
+
 # 2021-10-01 - v5.7.1-rc.2
 - Added additional iOS flags